+++ API/src/main/java/com/comze_instancelabs/minigamesapi/config/GunsConfig.java
@@ -17,6 +17,7 @@ package com.comze_instancelabs.minigamesapi.config;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
@@ -119,7 +120,7 @@ public class GunsConfig
         final InputStream defConfigStream = this.plugin.getResource("guns.yml");
         if (defConfigStream != null)
         {
-            final YamlConfiguration defConfig = YamlConfiguration.loadConfiguration(defConfigStream);
+            final YamlConfiguration defConfig = YamlConfiguration.loadConfiguration(new InputStreamReader(defConfigStream));
             this.arenaConfig.setDefaults(defConfig);
         }
     }