+++ API/src/main/java/com/comze_instancelabs/minigamesapi/config/HologramsConfig.java
@@ -17,6 +17,7 @@ package com.comze_instancelabs.minigamesapi.config;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
@@ -76,7 +77,7 @@ public class HologramsConfig
         final InputStream defConfigStream = this.plugin.getResource("holograms.yml");
         if (defConfigStream != null)
         {
-            final YamlConfiguration defConfig = YamlConfiguration.loadConfiguration(defConfigStream);
+            final YamlConfiguration defConfig = YamlConfiguration.loadConfiguration(new InputStreamReader(defConfigStream));
             this.holoConfig.setDefaults(defConfig);
         }
     }