+++ API/src/main/java/com/comze_instancelabs/minigamesapi/config/MessagesConfig.java
@@ -17,6 +17,7 @@ package com.comze_instancelabs.minigamesapi.config;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.Map;
 
@@ -365,7 +366,7 @@ public class MessagesConfig
         final InputStream defConfigStream = this.plugin.getResource("messages.yml");
         if (defConfigStream != null)
         {
-            final YamlConfiguration defConfig = YamlConfiguration.loadConfiguration(defConfigStream);
+            final YamlConfiguration defConfig = YamlConfiguration.loadConfiguration(new InputStreamReader(defConfigStream));
             this.messagesConfig.setDefaults(defConfig);
         }
     }