+++ API/src/main/java/com/comze_instancelabs/minigamesapi/config/PartyMessagesConfig.java
@@ -17,6 +17,7 @@ package com.comze_instancelabs.minigamesapi.config;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 
 import org.bukkit.ChatColor;
 import org.bukkit.configuration.file.FileConfiguration;
@@ -123,7 +124,7 @@ public class PartyMessagesConfig
         final InputStream defConfigStream = this.plugin.getResource("partymessages.yml");
         if (defConfigStream != null)
         {
-            final YamlConfiguration defConfig = YamlConfiguration.loadConfiguration(defConfigStream);
+            final YamlConfiguration defConfig = YamlConfiguration.loadConfiguration(new InputStreamReader(defConfigStream));
             this.messagesConfig.setDefaults(defConfig);
         }
     }