+++ Spigot-Test/src/main/java/com/comze_instancelabs/minigamesapi/spigottest/DummyServer.java
@@ -24,13 +24,13 @@ import org.bukkit.World;
 import org.bukkit.World.Environment;
 import org.bukkit.WorldType;
 import org.bukkit.configuration.file.YamlConfiguration;
-import org.bukkit.craftbukkit.v1_11_R1.CraftServer;
-import org.bukkit.craftbukkit.v1_11_R1.CraftWorld;
-import org.bukkit.craftbukkit.v1_11_R1.inventory.CraftItemFactory;
-import org.bukkit.craftbukkit.v1_11_R1.scheduler.CraftScheduler;
-import org.bukkit.craftbukkit.v1_11_R1.scoreboard.CraftScoreboard;
-import org.bukkit.craftbukkit.v1_11_R1.scoreboard.CraftScoreboardManager;
-import org.bukkit.craftbukkit.v1_11_R1.util.Versioning;
+import org.bukkit.craftbukkit.v1_12_R1.CraftServer;
+import org.bukkit.craftbukkit.v1_12_R1.CraftWorld;
+import org.bukkit.craftbukkit.v1_12_R1.inventory.CraftItemFactory;
+import org.bukkit.craftbukkit.v1_12_R1.scheduler.CraftScheduler;
+import org.bukkit.craftbukkit.v1_12_R1.scoreboard.CraftScoreboard;
+import org.bukkit.craftbukkit.v1_12_R1.scoreboard.CraftScoreboardManager;
+import org.bukkit.craftbukkit.v1_12_R1.util.Versioning;
 import org.bukkit.entity.Player;
 import org.bukkit.generator.ChunkGenerator;
 import org.bukkit.plugin.Plugin;
@@ -39,22 +39,22 @@ import org.bukkit.scoreboard.Team;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 
-import net.minecraft.server.v1_11_R1.Chunk;
-import net.minecraft.server.v1_11_R1.DedicatedPlayerList;
-import net.minecraft.server.v1_11_R1.DefinedStructureManager;
-import net.minecraft.server.v1_11_R1.EnumGamemode;
-import net.minecraft.server.v1_11_R1.ExceptionWorldConflict;
-import net.minecraft.server.v1_11_R1.IChunkLoader;
-import net.minecraft.server.v1_11_R1.IDataManager;
-import net.minecraft.server.v1_11_R1.IPlayerFileData;
-import net.minecraft.server.v1_11_R1.MethodProfiler;
-import net.minecraft.server.v1_11_R1.MinecraftServer;
-import net.minecraft.server.v1_11_R1.NBTTagCompound;
-import net.minecraft.server.v1_11_R1.PlayerChunkMap;
-import net.minecraft.server.v1_11_R1.WorldData;
-import net.minecraft.server.v1_11_R1.WorldProvider;
-import net.minecraft.server.v1_11_R1.WorldServer;
-import net.minecraft.server.v1_11_R1.WorldSettings;
+import net.minecraft.server.v1_12_R1.Chunk;
+import net.minecraft.server.v1_12_R1.DedicatedPlayerList;
+import net.minecraft.server.v1_12_R1.DefinedStructureManager;
+import net.minecraft.server.v1_12_R1.EnumGamemode;
+import net.minecraft.server.v1_12_R1.ExceptionWorldConflict;
+import net.minecraft.server.v1_12_R1.IChunkLoader;
+import net.minecraft.server.v1_12_R1.IDataManager;
+import net.minecraft.server.v1_12_R1.IPlayerFileData;
+import net.minecraft.server.v1_12_R1.MethodProfiler;
+import net.minecraft.server.v1_12_R1.MinecraftServer;
+import net.minecraft.server.v1_12_R1.NBTTagCompound;
+import net.minecraft.server.v1_12_R1.PlayerChunkMap;
+import net.minecraft.server.v1_12_R1.WorldData;
+import net.minecraft.server.v1_12_R1.WorldProvider;
+import net.minecraft.server.v1_12_R1.WorldServer;
+import net.minecraft.server.v1_12_R1.WorldSettings;
 
 /**
  * originally taken from spigot test sources
@@ -492,27 +492,27 @@ class DummyServer implements InvocationHandler, Answer {
             final UUID uuid = UUID.randomUUID();
             final ChunkGenerator generator = null;
             int dimension = this.worlds.size() + 1;
-            net.minecraft.server.v1_11_R1.WorldType mctype = null;
+            net.minecraft.server.v1_12_R1.WorldType mctype = null;
             switch (type)
             {
                 case AMPLIFIED:
-                    mctype = net.minecraft.server.v1_11_R1.WorldType.AMPLIFIED;
+                    mctype = net.minecraft.server.v1_12_R1.WorldType.AMPLIFIED;
                     break;
                 case CUSTOMIZED:
-                    mctype = net.minecraft.server.v1_11_R1.WorldType.CUSTOMIZED;
+                    mctype = net.minecraft.server.v1_12_R1.WorldType.CUSTOMIZED;
                     break;
                 default:
                 case FLAT:
-                    mctype = net.minecraft.server.v1_11_R1.WorldType.FLAT;
+                    mctype = net.minecraft.server.v1_12_R1.WorldType.FLAT;
                     break;
                 case LARGE_BIOMES:
-                    mctype = net.minecraft.server.v1_11_R1.WorldType.LARGE_BIOMES;
+                    mctype = net.minecraft.server.v1_12_R1.WorldType.LARGE_BIOMES;
                     break;
                 case NORMAL:
-                    mctype = net.minecraft.server.v1_11_R1.WorldType.NORMAL;
+                    mctype = net.minecraft.server.v1_12_R1.WorldType.NORMAL;
                     break;
                 case VERSION_1_1:
-                    mctype = net.minecraft.server.v1_11_R1.WorldType.NORMAL_1_1;
+                    mctype = net.minecraft.server.v1_12_R1.WorldType.NORMAL_1_1;
                     break;
                 
             }
@@ -541,27 +541,21 @@ class DummyServer implements InvocationHandler, Answer {
                 {
                     return new IChunkLoader(){
                         @Override
-                        public Chunk a(net.minecraft.server.v1_11_R1.World arg0, int arg1, int arg2) throws IOException
+                        public Chunk a(net.minecraft.server.v1_12_R1.World arg0, int arg1, int arg2) throws IOException
                         {
                             final Map<Integer, Chunk> map = chunks.get(arg1);
                             return map == null ? null : map.get(arg2);
                         }
 
                         @Override
-                        public void a(net.minecraft.server.v1_11_R1.World arg0, Chunk arg1) throws IOException, ExceptionWorldConflict
+                        public void a(net.minecraft.server.v1_12_R1.World arg0, Chunk arg1) throws IOException, ExceptionWorldConflict
                         {
                             final Map<Integer, Chunk> map = chunks.computeIfAbsent(arg1.locX, (x) -> new HashMap<>());
                             map.put(arg1.locZ, arg1);
                         }
 
                         @Override
-                        public void b(net.minecraft.server.v1_11_R1.World arg0, Chunk arg1) throws IOException
-                        {
-                            // ignore
-                        }
-
-                        @Override
-                        public void a()
+                        public void b(net.minecraft.server.v1_12_R1.World arg0, Chunk arg1) throws IOException
                         {
                             // ignore
                         }
@@ -573,7 +567,14 @@ class DummyServer implements InvocationHandler, Answer {
                         }
 
                         @Override
-                        public boolean a(int paramInt1, int paramInt2)
+                        public void c()
+                        {
+                            // TODO Auto-generated method stub
+                            
+                        }
+
+                        @Override
+                        public boolean chunkExists(int arg0, int arg1)
                         {
                             // TODO Auto-generated method stub
                             return false;